some refactoring on UI
