+++ Tasks/Task 13/GoodWebSite/GoodWebSite.UI/src/services/userClaimsManager.ts
