+++ Tasks/Task 13/GoodWebSite/GoodWebSite.UI/src/stores/userStore.ts
@@ -1,4 +1,4 @@
-import {isAccessTokenExpired} from "../functions/accessTokenManager.ts";
+import {isAccessTokenExpired} from "../services/accessTokenManager.ts";
 import {makeAutoObservable} from "mobx";
 
 export class UserStore{
